Workaround for broke CI with YARD 0.9.20
This commit solves the following "The manual directory is out of sync"
error.

```console
% ruby -v
ruby 2.7.0dev (2019-12-11T07:52:06Z master 3098798044) [x86_64-darwin17]
% CI=true bundle exec rake

(snip)

The manual directory is out of sync. Run `rake
generate_cops_documentation` and commit the results.
```

https://circleci.com/gh/rubocop-hq/rubocop-rails/2180

This issue will be resolved with the release of YARD 0.9.21 or higher,
which includes the following changes:
lsegal/yard#1290